more refactoring out of tests
